lesson04-model-validation/pyproject.toml
[project]
name = "lesson04-model-validation"
version = "0.1.0"
description = "Lesson 04 model validation mini-application."
requires-python = ">=3.13,<3.14"
dependencies = [
    "numpy",
    "scikit-learn",
    "azure-ai-projects",
    "azure-identity",
    "python-dotenv",
]

[tool.uv]
package = false